+++ lib/e_menu/src/apave.tcl
@@ -7,7 +7,7 @@
 ###########################################################
 
 package require Tk
-package provide apave 4.4.0
+package provide apave 4.4.1
 
 source [file join [file dirname [info script]] apavedialog.tcl]